Refactor project data and functions
