+++ src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
 import BaseLayout from "../layouts/BaseLayout.astro";
-import { getFeaturedProjects, type Project } from "../data/projects";
+import { getFeaturedProjects } from "../data/projectFunctions";
 
 const pageTitle = "Welcome!";
